requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0,<4.0
psycopg2-binary>=2.9.9,<3.0
celery>=5.3.0,<6.0
redis>=5.0.0,<6.0
pandas>=2.2.0,<3.0
openpyxl>=3.1.0,<4.0
